int64 change on preview
Squashed commit of the following:

commit 9c8ff52a7d47fd7f1ec59857db779385364f6a41
Merge: a367d13 0556f7a
Author: Jason Yin <jason.yin@oracle.com>
Date:   Tue Jul 10 19:41:51 2018 -0700

    Merge remote-tracking branch 'origin/preview' into jayin/regen

    # Conflicts:
    #	telemetry/telemetry_client.go

commit a367d1375fa5afa849893db1df98644d4fc5d981
Merge: 3b936f9 0ef0356
Author: Jason Yin <jason.yin@oracle.com>
Date:   Tue Jul 10 10:16:36 2018 -0700

    Merge remote-tracking branch 'origin/preview' into jayin/regen

    # Conflicts:
    #	telemetry/telemetry_client.go

commit 3b936f9acafb7b7c6008f018ec3a75f1e6a5a1fb
Author: Jason Yin <jason.yin@oracle.com>
Date:   Mon Jul 9 21:47:47 2018 -0700

    update object storage tests

commit ba4b0b76c957b95dc4333f46e2434b099bb4c56d
Author: Jason Yin <jason.yin@oracle.com>
Date:   Mon Jul 9 21:24:14 2018 -0700

    update db test cases

commit 2d9787bb7af115984c99f1adc19c6f8cb6404e66
Author: Jason Yin <jason.yin@oracle.com>
Date:   Mon Jul 9 15:08:32 2018 -0700

    update test cases to int64

commit c724b396ff02f0f2b3ae11f104b07e08105e5cbc
Author: Jason Yin <jason.yin@oracle.com>
Date:   Mon Jul 9 14:39:30 2018 -0700

    re-gen preview with latest codegen

+++ core/boot_volume_backup.go
@@ -63,7 +63,7 @@ type BootVolumeBackup struct {
 	ImageId *string `mandatory:"false" json:"imageId"`
 
 	// The size of the boot volume, in GBs.
-	SizeInGBs *int `mandatory:"false" json:"sizeInGBs"`
+	SizeInGBs *int64 `mandatory:"false" json:"sizeInGBs"`
 
 	// The OCID of the source boot volume backup.
 	SourceBootVolumeBackupId *string `mandatory:"false" json:"sourceBootVolumeBackupId"`
@@ -79,7 +79,7 @@ type BootVolumeBackup struct {
 
 	// The size used by the backup, in GBs. It is typically smaller than sizeInGBs, depending on the space
 	// consumed on the boot volume and whether the backup is full or incremental.
-	UniqueSizeInGBs *int `mandatory:"false" json:"uniqueSizeInGBs"`
+	UniqueSizeInGBs *int64 `mandatory:"false" json:"uniqueSizeInGBs"`
 }
 
 func (m BootVolumeBackup) String() string {
