+++ core/create_volume_details.go
@@ -46,11 +46,11 @@ type CreateVolumeDetails struct {
 	KmsKeyId *string `mandatory:"false" json:"kmsKeyId"`
 
 	// The size of the volume in GBs.
-	SizeInGBs *int `mandatory:"false" json:"sizeInGBs"`
+	SizeInGBs *int64 `mandatory:"false" json:"sizeInGBs"`
 
 	// The size of the volume in MBs. The value must be a multiple of 1024.
 	// This field is deprecated. Use sizeInGBs instead.
-	SizeInMBs *int `mandatory:"false" json:"sizeInMBs"`
+	SizeInMBs *int64 `mandatory:"false" json:"sizeInMBs"`
 
 	// Specifies the volume source details for a new Block volume. The volume source is either another Block volume in the same Availability Domain or a Block volume backup.
 	// This is an optional field. If not specified or set to null, the new Block volume will be empty.
@@ -75,8 +75,8 @@ func (m *CreateVolumeDetails) UnmarshalJSON(data []byte) (e error) {
 		DisplayName        *string                           `json:"displayName"`
 		FreeformTags       map[string]string                 `json:"freeformTags"`
 		KmsKeyId           *string                           `json:"kmsKeyId"`
-		SizeInGBs          *int                              `json:"sizeInGBs"`
-		SizeInMBs          *int                              `json:"sizeInMBs"`
+		SizeInGBs          *int64                            `json:"sizeInGBs"`
+		SizeInMBs          *int64                            `json:"sizeInMBs"`
 		SourceDetails      volumesourcedetails               `json:"sourceDetails"`
 		VolumeBackupId     *string                           `json:"volumeBackupId"`
 		AvailabilityDomain *string                           `json:"availabilityDomain"`
