+++ core/update_boot_volume_details.go
@@ -31,7 +31,7 @@ type UpdateBootVolumeDetails struct {
 	FreeformTags map[string]string `mandatory:"false" json:"freeformTags"`
 
 	// The size to resize the volume to in GBs. Has to be larger than the current size.
-	SizeInGBs *int `mandatory:"false" json:"sizeInGBs"`
+	SizeInGBs *int64 `mandatory:"false" json:"sizeInGBs"`
 }
 
 func (m UpdateBootVolumeDetails) String() string {
